refactor(S3): S3 디렉토리 경로 이동
