+++ src/main/java/com/matchday/matchdayserver/team/controller/TeamController.java
@@ -1,8 +1,8 @@
 package com.matchday.matchdayserver.team.controller;
 
 import com.matchday.matchdayserver.common.response.ApiResponse;
-import com.matchday.matchdayserver.common.s3.Service.S3PresignedService;
-import com.matchday.matchdayserver.common.s3.dto.S3PresignedResponse;
+import com.matchday.matchdayserver.s3.Service.S3PresignedService;
+import com.matchday.matchdayserver.s3.dto.S3PresignedResponse;
 import com.matchday.matchdayserver.team.model.dto.request.TeamCreateRequest;
 import com.matchday.matchdayserver.team.model.dto.response.TeamResponse;
 import com.matchday.matchdayserver.team.model.dto.response.TeamMemberListResponse;
